Added UDS routine to reset BOOT0 option bytes.
Added routine to salvage a board on which the wrong firmware was installed, by allowing user to reset the BOOT0 option byte of an ECU with a UDS command. This is mainly for when a user wrongly flashes ECU A firmware to ECU B/C/D and the ECUs now refuse to go into bootloader mode.