Added feature to let ECU A include slcan messages in its DBC processing.
This allows updating the values shown on screen when they are sent over slcan (and are technically not received by ECU A from CAN, since ECU A is the sender).
This does not cover UDS messages and answer to RTR frames.
Note that ECU A may update the values on screen slightly before the corresponding messages are sent on the CAN bus.