OOB in SP variable
SP is an index to the stack variable, which is made up of 32 uint16_t elements. The check here ensured that SP was not greater than the size of the stack variable; however, sizeof() will return the total size in bytes of the variable, not the number of elements.

This was changed to (sizeof(stack)/sizeof(uint16_t)) so it would return the correct number of elements.